requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.24.3
pyodbc==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
gunicorn==21.2.0
werkzeug==3.0.1